+++ doorstop/core/test/test_tree.py
@@ -12,7 +12,7 @@ import tempfile
 import operator
 import logging
 
-from doorstop.common import DoorstopError
+from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
 from doorstop.core.tree import Tree, build
 from doorstop.core.document import Document
 
@@ -100,7 +100,7 @@ class TestTreeStrings(unittest.TestCase):  # pylint: disable=R0904
         docs = [a, b, c]
         tree = Tree.from_list(docs)
         self.assertEqual(3, len(tree))
-        tree.check()
+        self.assertTrue(tree.check())
 
     def test_from_list_no_root(self):
         """Verify an error occurs when the tree has no root."""
@@ -144,18 +144,26 @@ class TestTree(unittest.TestCase):  # pylint: disable=R0904
                                parent='REQ')
         self.assertRaises(DoorstopError, tree._place, doc)  # pylint: disable=W0212
 
-    @patch('doorstop.core.document.Document.check')
-    def test_check(self, mock_check):
+    @patch('doorstop.core.document.Document.iter_issues')
+    def test_check(self, mock_iter_issues):
         """Verify trees can be checked."""
         logging.info("tree: {}".format(self.tree))
-        self.tree.check()
-        self.assertEqual(2, mock_check.call_count)
+        self.assertTrue(self.tree.check())
+        self.assertEqual(2, mock_iter_issues.call_count)
 
     @unittest.skipUnless(os.getenv(ENV), REASON)
     def test_check_long(self):
         """Verify trees can be checked (long)."""
         logging.info("tree: {}".format(self.tree))
-        self.tree.check()
+        self.assertTrue(self.tree.check())
+
+    def test_check_error(self):
+        """Verify an document error fails the tree check."""
+        with patch.object(self.tree, 'iter_issues',
+                          Mock(return_value=[DoorstopError('e'),
+                                             DoorstopWarning('w'),
+                                             DoorstopInfo('i')])):
+            self.assertFalse(self.tree.check())
 
     def test_new(self):
         """Verify a new document can be created on a tree."""